unify build scripts, update cheatcodes build script, fix vm.getFoundryVersion() cheatcode
