+++ crates/anvil/Cargo.toml
@@ -18,14 +18,6 @@ name = "anvil"
 path = "src/anvil.rs"
 required-features = ["cli"]
 
-[build-dependencies]
-foundry-common.workspace = true
-vergen = { workspace = true, default-features = false, features = [
-    "build",
-    "git",
-    "gitcl",
-] }
-
 [dependencies]
 # foundry internal
 anvil-core = { path = "core", features = ["serde", "impersonated-tx"] }
