+++ crates/cast/Cargo.toml
@@ -20,14 +20,6 @@ name = "cast"
 name = "cast"
 path = "bin/main.rs"
 
-[build-dependencies]
-foundry-common.workspace = true
-vergen = { workspace = true, default-features = false, features = [
-    "build",
-    "git",
-    "gitcl",
-] }
-
 [dependencies]
 # lib
 foundry-block-explorers.workspace = true
