+++ crates/cast/bin/args.rs
@@ -9,11 +9,12 @@ use alloy_primitives::{Address, B256, U256};
 use alloy_rpc_types::BlockId;
 use clap::{Parser, Subcommand, ValueHint};
 use eyre::Result;
-use foundry_cli::opts::{EtherscanOpts, GlobalArgs, RpcOpts};
+use foundry_cli::{
+    opts::{EtherscanOpts, GlobalArgs, RpcOpts},
+    VERSION_MESSAGE,
+};
 use foundry_common::ens::NameOrAddress;
 use std::{path::PathBuf, str::FromStr};
-
-const VERSION_MESSAGE: &str = env!("FOUNDRY_SHORT_VERSION");
 
 /// Perform Ethereum RPC calls from the comfort of your command line.
 #[derive(Parser)]
