+++ crates/cheatcodes/Cargo.toml
@@ -14,13 +14,6 @@ exclude.workspace = true
 [lints]
 workspace = true
 
-[build-dependencies]
-vergen = { workspace = true, default-features = false, features = [
-    "build",
-    "git",
-    "gitcl",
-] }
-
 [dependencies]
 foundry-cheatcodes-spec.workspace = true
 foundry-common.workspace = true
@@ -49,7 +42,6 @@ alloy-network.workspace = true
 alloy-rlp.workspace = true
 
 base64.workspace = true
-chrono.workspace = true
 dialoguer = "0.11"
 eyre.workspace = true
 itertools.workspace = true
@@ -69,3 +61,10 @@ tracing.workspace = true
 walkdir.workspace = true
 proptest.workspace = true
 serde.workspace = true
+
+[build-dependencies]
+vergen = { workspace = true, default-features = false, features = [
+    "build",
+    "git",
+    "gitcl",
+] }
