+++ crates/chisel/Cargo.toml
@@ -20,14 +20,6 @@ workspace = true
 name = "chisel"
 path = "bin/main.rs"
 
-[build-dependencies]
-foundry-common.workspace = true
-vergen = { workspace = true, default-features = false, features = [
-    "build",
-    "git",
-    "gitcl",
-] }
-
 [dependencies]
 # forge
 forge-fmt.workspace = true
