+++ crates/chisel/bin/main.rs
@@ -13,6 +13,7 @@ use foundry_cli::{
     handler,
     opts::{BuildOpts, GlobalArgs},
     utils::{self, LoadConfig},
+    VERSION_MESSAGE,
 };
 use foundry_common::{evm::EvmArgs, fs};
 use foundry_config::{
@@ -36,8 +37,6 @@ static ALLOC: tikv_jemallocator::Jemalloc = tikv_jemallocator::Jemalloc;
 
 // Loads project's figment and merges the build cli arguments into it
 foundry_config::merge_impl_figment_convert!(Chisel, build, evm);
-
-const VERSION_MESSAGE: &str = env!("FOUNDRY_SHORT_VERSION");
 
 /// Fast, utilitarian, and verbose Solidity REPL.
 #[derive(Debug, Parser)]
