+++ crates/cli/Cargo.toml
@@ -1,6 +1,5 @@
 [package]
 name = "foundry-cli"
-
 version.workspace = true
 edition.workspace = true
 rust-version.workspace = true
@@ -8,6 +7,7 @@ authors.workspace = true
 license.workspace = true
 homepage.workspace = true
 repository.workspace = true
+build = "build.rs"
 
 [lints]
 workspace = true
@@ -50,16 +50,15 @@ yansi.workspace = true
 
 tracing-tracy = { version = "0.11", optional = true }
 
-[dev-dependencies]
-tempfile.workspace = true
-
 [build-dependencies]
-foundry-common.workspace = true
 vergen = { workspace = true, default-features = false, features = [
     "build",
     "git",
     "gitcl",
 ] }
+
+[dev-dependencies]
+tempfile.workspace = true
 
 [features]
 default = ["rustls"]
