+++ crates/cli/src/opts/global.rs
@@ -1,9 +1,8 @@
 use clap::{ArgAction, Parser};
-use foundry_common::{
-    shell::{ColorChoice, OutputFormat, OutputMode, Shell, Verbosity},
-    version::NIGHTLY_VERSION_WARNING_MESSAGE,
-};
+use foundry_common::shell::{ColorChoice, OutputFormat, OutputMode, Shell, Verbosity};
 use serde::{Deserialize, Serialize};
+
+use crate::NIGHTLY_VERSION_WARNING_MESSAGE;
 
 /// Global arguments for the CLI.
 #[derive(Clone, Debug, Default, Serialize, Deserialize, Parser)]
