+++ crates/common/Cargo.toml
@@ -69,13 +69,6 @@ anstream.workspace = true
 anstyle.workspace = true
 terminal_size.workspace = true
 
-[build-dependencies]
-vergen = { workspace = true, default-features = false, features = [
-    "build",
-    "git",
-    "gitcl",
-] }
-
 [dev-dependencies]
 foundry-macros.workspace = true
 similar-asserts.workspace = true
