+++ crates/forge/bin/opts.rs
@@ -8,10 +8,8 @@ use crate::cmd::{
 use clap::{Parser, Subcommand, ValueHint};
 use forge_script::ScriptArgs;
 use forge_verify::{VerifyArgs, VerifyBytecodeArgs, VerifyCheckArgs};
-use foundry_cli::opts::GlobalArgs;
+use foundry_cli::{opts::GlobalArgs, VERSION_MESSAGE};
 use std::path::PathBuf;
-
-const VERSION_MESSAGE: &str = env!("FOUNDRY_SHORT_VERSION");
 
 /// Build, test, fuzz, debug and deploy Solidity contracts.
 #[derive(Parser)]
